factor directives into modules
